+++ bin/actions/0220022-50.sh
@@ -1,12 +1,23 @@
 #!/bin/bash
 
-#: ◣ + M - Launch or bring Claws Mail to front
+#: ◣ + M -
+# Matterost desktop client: Open "direct messages" popup
+# Else: Launch or bring Claws Mail to front
 
 focusApplication=\
 $(cat /proc/"$(xdotool getwindowpid "$(xdotool getwindowfocus)")"/comm)
 
 path_self="$( cd "$(dirname "$0")" >/dev/null 2>&1 || exit ; pwd -P )"
 "$path_self"/utils/remove_unwanted_output.sh "$focusApplication"
+
+if [[ "$focusApplication" =~ "mattermost-desk" ]]; then
+  xdotool sleep 0.1
+  xdotool key Ctrl+Shift+k
+
+  unset path_self
+  unset focusApplication
+  exit 0
+fi
 
 if pidof -s claws-mail > /dev/null; then
 	wmctrl -a "Claws Mail"